feat: extract db scaffolding logic and update Keycloak CLI image version
